moved traverseUserWorkspace funcionality into the traverseParsingFunctions and renamed the optimizeFunctions file to dynamicImportFunctions
